Add files via upload
This C++ program defines a Rectangle class with private attributes length and breadth, a constructor, and a member function area() that calculates the area of the rectangle.

In the main function, an instance of the Rectangle class is created with a length of 10 and a breadth of 5. The area() function is then called to compute and print the area of the rectangle.